requirements.txt
accelerate==1.2.1
attrs==24.3.0
av==14.1.0
datasets
deepspeed==0.16.3
#flash-attn==2.7.2.post1
huggingface-hub==0.28.1
hydra-core
numpy==1.26.4
nvitop
omegaconf
openai
opencv-python
opencv-python-headless
peft==0.14.0
pillow==10.4.0
qwen-vl-utils
ray
rich
sentencepiece==0.2.0
tokenizers==0.21.0
torch
torchaudio
torchvision
tqdm
transformers
trl
vllm
wandb
gcsfs==2024.12.0
fsspec==2024.12.0